pyproject.toml
[tool.poetry]
name = "pdca"
version = "0.1.0"
description = ""
authors = ["karintou"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
poethepoet = "^0.24.4"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
asyncpg = "^0.29.0"
alembic = "^1.12.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.0"
sqlalchemy-utils = "^0.41.1"
psycopg2 = "^2.9.9"


[tool.poetry.group.dev.dependencies]
mypy = "^1.7.0"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
pyproject-flake8 = "^6.1.0"
pytest-asyncio = "^0.23.2"
httpx = "^0.25.2"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.poe.tasks]
lint = "pflake8 ."
format = "bash -c 'isort . && black .'"
format-check = "bash -c 'isort --check . && black --check .'"
type = "mypy main.py api/"

[tool.mypy]
strict = true
ignore_missing_imports = true
plugins = "sqlalchemy.ext.mypy.plugin"

[tool.flake8]
exclude = [".venv", ".git", "__pycache__",]
max-complexity = 10
max-line-length = 88